requirements.txt
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
parsel==1.5.2
Pillow==7.1.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pytz==2019.3
queuelib==1.5.0
requests==2.23.0
Scrapy==2.1.0
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
zope.interface==5.1.0
